+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to
 ## [Unreleased]
 ### Added
 - `ff show` command to show specific id
-- Proper error messeges, when id not found.
+- Proper error messages, when id not found.
 
 ### Changed
 - UI: display UUIDs without quotes